some changes to accomodate clang cuda on Mac, some refactoring of task scheduler to make space for changes
