+++ cmake/kokkos_build.cmake
@@ -76,7 +76,7 @@ IF(KOKKOS_SEPARATE_LIBS)
   )
 
   foreach(lib IN LISTS KOKKOS_TPL_LIBRARY_NAMES)
-    if ("${lib}" STREQUAL "cuda")
+    if (("${lib}" STREQUAL "cuda") AND (NOT "${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang"))
       set(LIB_cuda "-lcuda")
     else()
       find_library(LIB_${lib} ${lib} PATHS ${KOKKOS_TPL_LIBRARY_DIRS})
@@ -158,7 +158,7 @@ ELSE()
   )
 
   foreach(lib IN LISTS KOKKOS_TPL_LIBRARY_NAMES)
-    if ("${lib}" STREQUAL "cuda")
+    if (("${lib}" STREQUAL "cuda") AND (NOT "${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang"))
       set(LIB_cuda "-lcuda")
     else()
       find_library(LIB_${lib} ${lib} PATHS ${KOKKOS_TPL_LIBRARY_DIRS})
