+++ cmake/tribits.cmake
@@ -309,6 +309,10 @@ ENDFUNCTION()
 FUNCTION(TRIBITS_TPL_TENTATIVELY_ENABLE)
 ENDFUNCTION()
 
+FUNCTION(TRIBITS_ADD_ADVANCED_TEST)
+  # TODO Write this
+ENDFUNCTION()
+
 FUNCTION(TRIBITS_ADD_EXECUTABLE_AND_TEST EXE_NAME)
 
   SET(options STANDARD_PASS_OUTPUT WILL_FAIL)
