+++ core/perf_test/CMakeLists.txt
@@ -47,6 +47,7 @@ TRIBITS_ADD_EXECUTABLE(
   PerformanceTest_TaskDAG
   SOURCES test_taskdag.cpp
   COMM serial mpi
+  TESTONLYLIBS kokkos_gtest ${TEST_LINK_TARGETS}
   )
 
 TRIBITS_ADD_TEST(
